+++ src/HalKit/HalKit.csproj
@@ -35,17 +35,20 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="HalClient.cs" />
+    <Compile Include="Models\Response\CurieLink.cs" />
+    <Compile Include="Models\Response\Link.cs" />
+    <Compile Include="Models\Response\LinkCollection.cs" />
+    <Compile Include="Models\Response\Resource.cs" />
+    <Compile Include="Models\Response\RootResource.cs" />
     <Compile Include="Services\ILinkResolver.cs" />
     <Compile Include="Services\LinkResolver.cs" />
     <Compile Include="Http\HttpConnection.cs" />
     <Compile Include="Http\IHttpConnection.cs" />
     <Compile Include="IHalClient.cs" />
-    <Compile Include="Models\CurieLink.cs" />
     <Compile Include="Exceptions\LinkNotFoundException.cs" />
     <Compile Include="Json\DefaultJsonSerializer.cs" />
     <Compile Include="Json\EmbeddedAttribute.cs" />
     <Compile Include="Json\ResourceConverter.cs" />
-    <Compile Include="Models\LinkCollection.cs" />
     <Compile Include="Extensions\TaskExtensions.cs" />
     <Compile Include="HalKitConfiguration.cs" />
     <Compile Include="Http\ApiResponse.cs" />
@@ -55,15 +58,12 @@
     <Compile Include="Http\IHttpClientFactory.cs" />
     <Compile Include="IHalKitConfiguration.cs" />
     <Compile Include="Json\IJsonSerializer.cs" />
-    <Compile Include="Models\Link.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="..\..\SolutionInfo.cs">
       <Link>Properties\SolutionInfo.cs</Link>
     </Compile>
     <Compile Include="Http\HttpClientFactory.cs" />
     <Compile Include="Requires.cs" />
-    <Compile Include="Resources\Resource.cs" />
-    <Compile Include="Resources\RootResource.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <ItemGroup>
