+++ src/HalKit/Services/ILinkResolver.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using HalKit.Models;
+using HalKit.Models.Response;
 
 namespace HalKit.Services
 {
