+++ src/HalKit/Services/LinkResolver.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using HalKit.Models;
+using HalKit.Models.Response;
 using Tavis.UriTemplates;
 
 namespace HalKit.Services
