+++ tests/HalKit.Tests/Models/LinkCollectionTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using HalKit.Exceptions;
 using HalKit.Models;
+using HalKit.Models.Response;
 using Xunit;
 
 namespace HalKit.Tests.Models
