+++ tests/HalKit.Tests/Services/LinkResolverTests.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using HalKit.Models;
+using HalKit.Models.Response;
 using HalKit.Services;
 using Xunit;
